server/requirements.txt
# --- Web Server & API ---
fastapi
uvicorn[standard]
python-multipart

# --- Core AI/ML Models & Logic ---
google-generativeai         # For OCR with Gemini
sentence-transformers       # For Semantic Analysis (uses PyTorch)
torch>=2.0.0                # The backend for sentence-transformers
transformers                # A dependency of sentence-transformers
scikit-learn                # For Thematic Analysis
nltk                        # For Thematic Analysis tokenization

# --- Utilities & File Handling ---
python-dotenv               # For loading API keys from .env file
pdf2image                    # To convert PDF pages to images for OCR
Pillow                      # A dependency for pdf2image
numpy                       # A core dependency for most ML libraries